requirements.txt
flask
gunicorn
psycopg2-binary
werkzeug
jinja2
python-dotenv
bcrypt

